Improved Relation querying methods
This patch provides imporoved Relation querying methods.
